Cleanup create
